requirements.txt
flask==3.0.3
gunicorn==22.0.0
tensorflow>=2.16.0
opencv-python-headless>=4.9.0
Pillow>=10.3.0
numpy>=1.26.0
requests>=2.32.0
 